actions pineadas por SHA, permisos por job y fixes de aritmética entera en el bot

+++ src/main/java/com/puckzone/game/bot/BotBrain.java
@@ -154,7 +154,7 @@ public class BotBrain {
     }
 
     private double maxY() {
-        return props.boardHeight() - props.puckRadius();
+        return (double) props.boardHeight() - props.puckRadius();
     }
 
     private static double stepTowards(double current, double target, double maxStep) {
